Change role bucket restriction to bucket name
Blazer only creates application keys with bucket/path restrictions
by looking up the bucket, and only looks up the bucket by name.
Switch to using that, instead of the ID, as the restriction.

B2 documentation leads me to believe that not only is the Bucket
ID unique, but the bucket name will be globally unique with no
ability to reuse bucket names.